Decouple property computation from aggregation [Resolves #125] (#140)
A lasting problem has been the fragility of large aggregation jobs, that tend to require the coordination of CPU or memory-intensive jobs. Traditionally we have only save the aggregated results, which prevents us from reusing computation for different aggregations, or reusing a half-done list of computed properties, among many other problems.

Here we introduce The skills_ml.jon_postings.computed_properties module, which is split into two parts: computers (computing properties on collections), and aggregators (aggregating them).

The computers save data indexed by job posting to daily-partitioned storage on s3. This is cached per job posting id, so if a large collection craps out halfway through, the computation can be reused.

In addition, I'm trying to get the interface between Airflow and skills-ml right. I want to put as little code in Airflow as possible, in large part because it's very tough to unit test anything in Airflow. Just the same, we do want to be able to configure aggregations in Airflow. If we decide that now we want to create a new tabular dataset that mixes properties together in a different way, or increase the # of top skills that are present, this should be a simple change in Airflow. The result is a bit of a complex dance between the JobPostingPropertyComputers and their aggregators: the Computers define compatible aggregator functions that work with their output, and this is enforced by unit test. But aggregators can choose which ones they use for a particular aggregation (should be decided in Airflow).

To wrap up, there is now an example of a basic computation and aggregation task, runnable without any dependencies, that is basically a mini versionof the Data@Work Research Hub.

There was also a moto problem with the introduction of boto3, which I was actually able to fix by converting the geocoder and CBSA finders to the new S3BackedJsonDict class. They had implemented their own version using boto2, but I was able to remove a bunch of custom code and tests, as well as fix Travis, by switching them over.