Occupation Classifier pipeline components (#191)
I see the whole pipeline break down into Embedding, Classifier, Logger and Test. This PR is addressing the Embedding and Classifier.

CombinedClassifier: a class to combine embedding and classifier using scikit-learn pipeline object
OccupationClassifierTrainer: a trainer to train and cross-validate a series of classifier given the conifg file. Now the cross-validation result is the property, but we will want to convert it into a json format as well as the testing suite.
EmbeddingTransformer: scikit-learn interface for embedding model
create_training_set: a helper function to create training data
TrainingMatrix: a class holds training matrix, label matrix, embedding model and target variable to feed in a trainer
TargetVariable: a class has abstract property methods of name and transformer, and take self-defined filters as input.
SOCMajorGroup: is inherited from TargetVariable and has default properties of SocEncoder and two soc code filters(unknow_soc_filter and empty_soc_filter).