Skill Extraction Metrics [Resolves #146]
Skill extraction metrics are added. None of these are label-aware yet.
There are many other useful hypothesized metrics, but represents a base
to which more metrics can be added.

New modules:

- Introduce evaluation.skill_extractor_metrics, which introduces a base
class for any evaluation metrics, as well as several examples.
- Introduce evaluation.skill_extractors, which handles candidate skill
creation and metric computation given metrics

Refactoring to support the new modules:

- Add 'name' to Ontology, so a name can be used in the
OntologyOccupationRecall metric.
- Introduce get_onet_occupation, so we can move towards working with the
more standard occupationalCategory key in the job posting. Ideally we
should start using this more and move towards a future where the ETL
only populates standard fields.
- Update CandidateSkill to a namedtuple, include more needed contextual
fields in it so the original sample is not needed for evaluation, and add factory_boy to easily create large numbers of candidate skills in tests
- Update Sample to use Storage object and implement __len__

Documentation:

- Replace examples/UploadCandidateFromSample with
examples/SkillExtractionEvaluation, that runs the sample job postings
through exact and fuzzy matching with a variety of ONET subsets and
outputs a few metrics for each run.