Print Ontology Stats
Adds a print_summary_stats() method to the ontology class. In the
printing of these summary stats on ONET, some oddities were found with both the base class and the ONET loader that have been fixed here.

- Ontology name was not included in the serialized version, so if you
wrote and then loaded one it would lose the name
- ONET loader was being too permissive with regards to competency-occupation
edges. It was adding an edge for each line in the file (IM and LV),
which not only duplicated edges but also would add irrelevant edges
along with relevant ones. So instead, the loader only adds
competency-occupation edges from the IM (importance) row, and only if the importance
level is 3 or above. How was this arrived at? ONET's guidelines on the
site. Competencies are only rated as 'important' to the occupation and
displayed if they pass this threshold. This seems like a reasonable
guideline for our mapping.
- Base Competency and Occupation classes were not comparable. Added a
__lt__ and functools.total_ordering to allow a user to sort and group by
Competency and Occupation objects.