Extract Section Text [Resolves #216] (#217)
Introduces section extraction, which is aimed at zooming in on specific
sections of large pieces of unstructured text. It does this through
sentence/line/bullet heuristics. It comes in three flavors:

- basic utility function
- Word2Vec corpus creator that uses the utility function
- skill extractor that uses the utility function and assumes that each
sentence contained in the section is a distinct skill

The skill extractor is added to the skill extractor metric example,
along with some retooling so that non-list-based skill extractors make
more sense in that example.