Section Extract Fixes
In running the section extract skill extractor at scale, a couple of
problems were noticed:

- The encoding exception catch block in sentence_tokenize is always hit
- Empty strings break it entirely

Further experimenting with the encoding reveals that
nltk.sentence_tokenize seems to need a string, so I don't think we
should try and encode it at all.

Also, the section_extract had some oddities: it didn't remove any
bullets from returned sections. This is a very minor problem but
removing the bullets does improve the readability of the results

Fixes:

- Add empty document check for section_extract
- Remove encoding block in sentence_tokenize
- Remove bullets from section_extract output
- Remove commented out code in section extract test