Include real start index in section extract skill extractor [Resolves
To be as useful as the other skill extractors, SectionExtract needs to
report the start index of the match in each yielded CandidateSkill. To enable this, we add an include_span argument to the sentence tokenizer. Doing this actually helps us remove some index tracking code from other skill extractors.